requirements.txt
openai-whisper
pyaudio
pyperclip
pynput
pyyaml
pyinstaller
